+++ lib/concerns_on_rails/legacy_aliases.rb
@@ -14,4 +14,5 @@ module ConcernsOnRails
   Activatable   = Models::Activatable
   Tokenizable   = Models::Tokenizable
   Stateable     = Models::Stateable
+  Addressable   = Models::Addressable
 end